% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paris-bikes-package.R
\docType{package}
\name{parisbikes-package}
\alias{parisbikes}
\alias{parisbikes-package}
\title{parisbikes: What the Package Does}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
