% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{pb_centrality}
\alias{pb_centrality}
\title{Calculate centrality on a bicycle network}
\usage{
pb_centrality(network, elev_file, estimate = TRUE)
}
\arguments{
\item{network}{Full path to file produced by \link{pb_get_network} function,
as an Open Street Map network in 'silicate' (SC) format.}

\item{elev_file}{Local path to a 'geotiff' file of elevation data covering
the geographical area of the network, and downloaded as explained in the
\pkg{osmdata} function 'osm_elevation'.}

\item{estimate}{If `TRUE`, calculate an initial estimate of how long the
centrality calculation is likely to take.}
}
\value{
A \pkg{dodgr} 'street_network' object including a "centrality"
column.
}
\description{
Calculate centrality on a bicycle network
}
\note{
The centrality calculation is run in parallel, and can not be
interrupted once it has begun. It is recommended to use `estimate = TRUE`,
after which an interactive prompt will allow the calculation to proceed.
Calculations may only be stopped by killing the R process.
}
