% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-network.R
\name{pb_get_network}
\alias{pb_get_network}
\title{Get an OSM network for a nominated location in 'silicate' (SC) format}
\usage{
pb_get_network(location = "Paris France", path)
}
\arguments{
\item{location}{Name of location, passed to Nominatim server.}

\item{path}{Local path where network is to be saved. Final name will be the
first component of 'location', with '-sc.Rds' appended.}
}
\description{
This function is especially intended for large cities for which overpass will
generally not deliver the whole network in one dump, and so divides
everything up into the requisite sequences of key-value pairs, and
reassembles the entire network at the end.
}
